:boom: seriously rewrite the entire prefix check system
Let's talk about directory mount directories: they are uncreateable, because that is ***dumb***.

The current system was absolutely insensitive. It thought it instead was really good to create directory mount directories, but that is ***dumb***.

The new system is instead a hack that forces the user to manually create $PREFIX as an existing and empty directory. Only way to be sure, apparently.

I hope there'll be someone who can finally fix that outstanding bug.